Add files via upload
trying to work on a function that generates a list of the possible configurations of blocks in a given board. started with tiny_5 to understand the problem better. the configurations are being printed correctly but for some reason i can't save them to a list? maybe something weird about numpy data types?