adding new wip doc
Fixed the problem with the starting blocks. Fixed my mistake with the numpy indexing (x and y had to be flipped). Now it still breaks when there's a clear block and a reflect block sandwiched around the start of the laser because it reflects into itself infinitely and exceeds number of recursions. Don't know how to fix it.

At this point I'm still working only on tiny_5. Need to implement recursion into the get_configs function to iterate through the blocks in the list