Add files via upload
This directory is the input file directory, in which we can store the input files that are needed to be solved. For checking a custom .bff file, you can simply place the file in this directory and run the main code and it will automatically go through all the .bff files in the directory. No need to specify any name of the file in the main python file.